+++ src/main/java/framework/base/FrameworkInitialize.java
@@ -3,6 +3,7 @@ package framework.base;
 import framework.config.Settings;
 import io.github.bonigarcia.wdm.EdgeDriverManager;
 import io.github.bonigarcia.wdm.InternetExplorerDriverManager;
+import lombok.extern.slf4j.Slf4j;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.edge.EdgeDriver;
@@ -10,6 +11,7 @@ import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.ie.InternetExplorerDriver;
 import org.openqa.selenium.opera.OperaDriver;
 
+@Slf4j
 public class FrameworkInitialize {
 
     private WebDriver driver = null;
@@ -21,43 +23,49 @@ public class FrameworkInitialize {
 
         this.userDir = System.getProperty("user.dir");
         this.sB = new StringBuilder(userDir);
-        this.sB.append("/src/main/resources/drivers/");
+        this.sB.append("\\src\\main\\resources\\drivers\\");
 
         if (System.getProperty("os.name").contains("Windows")) {
-            this.sB.append("windows/");
+            this.sB.append("windows\\");
             makeWindowsDriver();
         } else {
-            this.sB.append("unix/");
+            this.sB.append("unix\\");
             makeUnixDriver();
         }
 
         DriverContext.setDriver(driver);
         DriverContext.browser = new Browser(driver);
+        log.debug("Framework starts driver: {}", sB.toString());
     }
 
     private void makeWindowsDriver() {
         switch (browserType) {
             case chrome: {
-                System.setProperty("webdriver.chrome.driver", this.sB.append("chromedriver.exe").toString());
+                this.sB.append("chromedriver.exe");
+                System.setProperty("webdriver.chrome.driver", this.sB.toString());
                 driver = new ChromeDriver();
                 break;
             }
             case firefox: {
-                System.setProperty("webdriver.gecko.driver", this.sB.append("geckodriver.exe").toString());
+                this.sB.append("geckodriver.exe");
+                System.setProperty("webdriver.gecko.driver", this.sB.toString());
                 driver = new FirefoxDriver();
                 break;
             }
             case opera: {
-                System.setProperty("webdriver.opera.driver", this.sB.append("operadriver.exe").toString());
+                this.sB.append("operadriver.exe");
+                System.setProperty("webdriver.opera.driver", this.sB.toString());
                 driver = new OperaDriver();
                 break;
             }
             case ie: {
+                this.sB.append("InternetExplorer.exe");
                 InternetExplorerDriverManager.getInstance().setup();
                 driver = new InternetExplorerDriver();
                 break;
             }
             case edge: {
+                this.sB.append("MicrosoftEdge.exe");
                 EdgeDriverManager.getInstance().setup();
                 driver = new EdgeDriver();
                 break;
@@ -68,17 +76,20 @@ public class FrameworkInitialize {
     private void makeUnixDriver() {
         switch (browserType) {
             case chrome: {
-                System.setProperty("webdriver.chrome.driver", this.sB.append("chromedriver").toString());
+                this.sB.append("chromedriver");
+                System.setProperty("webdriver.chrome.driver", this.sB.toString());
                 driver = new ChromeDriver();
                 break;
             }
             case firefox: {
-                System.setProperty("webdriver.gecko.driver", this.sB.append("geckodriver").toString());
+                this.sB.append("geckodriver");
+                System.setProperty("webdriver.gecko.driver", this.sB.toString());
                 driver = new FirefoxDriver();
                 break;
             }
             case opera: {
-                System.setProperty("webdriver.opera.driver", this.sB.append("operadriver").toString());
+                this.sB.append("operadriver");
+                System.setProperty("webdriver.opera.driver", this.sB.toString());
                 driver = new OperaDriver();
                 break;
             }