+++ src/main/java/framework/config/ConfigReader.java
@@ -2,11 +2,13 @@ package framework.config;
 
 import framework.base.BrowserType;
 import framework.tools.TimeTool;
+import lombok.extern.slf4j.Slf4j;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Properties;
 
+@Slf4j
 public class ConfigReader {
 
     public static void populateSettings() {
@@ -23,7 +25,7 @@ public class ConfigReader {
         try {
             p.load(inputStream);
         } catch (IOException e) {
-
+            log.error("Failure to read configuration file: {}", e);
         }
 
         Settings.URLsite = p.getProperty("URLsite");