+++ src/main/java/framework/tools/LogsTool.java
@@ -4,27 +4,32 @@ import com.cucumber.listener.ExtentProperties;
 import com.cucumber.listener.Reporter;
 import framework.base.DriverContext;
 import framework.config.Settings;
+import lombok.extern.slf4j.Slf4j;
 import org.openqa.selenium.Capabilities;
 import org.openqa.selenium.remote.RemoteWebDriver;
 
 import java.io.File;
 
+@Slf4j
 public class LogsTool {
 
     public static void setExtentReportsSettings() {
         ExtentProperties extentProperties = ExtentProperties.INSTANCE;
         Settings.LogsPath = "output/" + Settings.StartTestsTime + "/";
         extentProperties.setReportPath(Settings.LogsPath + Settings.LogFileName);
+        log.debug("Extentreport will be saved to: {}", Settings.LogsPath + Settings.LogFileName);
     }
 
     public static void setExtentReportsInfo() {
 
         Capabilities caps = ((RemoteWebDriver) DriverContext.driver).getCapabilities();
+        final String extentConfigPath = "src\\main\\resources\\config\\extent-config.xml";
 
         try {
-            Reporter.loadXMLConfig(new File("src/main/resources/config/extent-config.xml"));
+            Reporter.loadXMLConfig(new File(extentConfigPath));
+            log.debug("The ExtentReports configuration file has been read from: {}", extentConfigPath);
         } catch (Exception e) {
-
+            log.error("Failure to read extent-config.xml: {}", e);
         }
 
         Reporter.setSystemInfo("user", System.getProperty("user.name"));